Animation loop data, VOC parser
+ Animation loop configuration is now loaded from res file
+ Partial VOC parser. More than is needed for SS
+ Progress indicator to map import
+ Uses object id from link data

+++ SSImporter/Runtime/Interfaces/CircuitAccess.cs
@@ -17,7 +17,7 @@ namespace SystemShock.Interfaces {
 
         private void Start() {
             SystemShockObject ssObject;
-            if (ActionData.ObjectToTrigger != 0 && levelInfo.Objects.TryGetValue((uint)ActionData.ObjectToTrigger, out ssObject)) {
+            if (ActionData.ObjectToTrigger != 0 && levelInfo.Objects.TryGetValue(ActionData.ObjectToTrigger, out ssObject)) {
                 Target = ssObject.GetComponent<Triggerable>();
 
                 if (Target == null)
