+++ SSImporter/Runtime/Interfaces/KeyPad.cs
@@ -18,7 +18,7 @@ namespace SystemShock.Interfaces {
 
         private void Start() {
             SystemShockObject ssObject;
-            if (ActionData.ObjectToTrigger1 != 0 && levelInfo.Objects.TryGetValue((uint)ActionData.ObjectToTrigger1, out ssObject)) {
+            if (ActionData.ObjectToTrigger1 != 0 && levelInfo.Objects.TryGetValue(ActionData.ObjectToTrigger1, out ssObject)) {
                 Target1 = ssObject.GetComponent<Triggerable>();
 
                 if (Target1 == null)
@@ -27,7 +27,7 @@ namespace SystemShock.Interfaces {
                 Debug.Log("Tried to find object! " + ActionData.ObjectToTrigger1, this);
             }
 
-            if (ActionData.ObjectToTrigger2 != 0 && levelInfo.Objects.TryGetValue((uint)ActionData.ObjectToTrigger2, out ssObject)) {
+            if (ActionData.ObjectToTrigger2 != 0 && levelInfo.Objects.TryGetValue(ActionData.ObjectToTrigger2, out ssObject)) {
                 Target2 = ssObject.GetComponent<Triggerable>();
 
                 if (Target2 == null)
