+++ SSImporter/Runtime/TriggerActions/Destroy.cs
@@ -16,7 +16,7 @@ namespace SystemShock.TriggerActions {
         }
 
         private void Start() {
-            if (ActionData.ObjectId != 0 && !levelInfo.Objects.TryGetValue((uint)ActionData.ObjectId, out Target))
+            if (ActionData.ObjectId != 0 && !levelInfo.Objects.TryGetValue(ActionData.ObjectId, out Target))
                 Debug.Log("Tried to find object! " + ActionData.ObjectId, this);
         }
